pyproject.toml
[project]
name = "rotten-tomatoes"
version = "0.1.0"
description = "Kalshi Rotten Tomatoes Prediction Market Bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.1.0",
    "requests>=2.31.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "cryptography>=46.0.3",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.0",
    "tenacity>=8.2.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.0",
    "notebook>=7.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
